Coverage / C-metric of two (approximation) sets

Given two (approximation) sets \(X = \{x_1, \ldots, x_{|X|}\}\) and \(Y = \{y_1, \ldots, y_{|Y|}\}\) this indidcator calculates the fraction of points from \(Y\) which are dominated by at least one point \(x \in X\). Formally it is defined as: $$ C(X, Y) = \frac{|\{y \in Y \mid \exists x \in X \, with \, x \preceq y\}|}{|Y|}. $$ It holds that \(C(X, Y) \in [0,1]\). Note that in general \(C(X, Y) \neq 1 - C(Y, X)\).

cov(x, y, ...)

Arguments

x
[matrix]
First point set in column major format.

y
[matrix]
Second point set in column major format.

...
[any]
Not used.

Value

Coverage value.

References

[1] E. Zitzler, Evolutionary algorithms for multiobjective optimization: Methods and applications, Ph.D. thesis, Swiss Federal Institute of Technology Zurich (1999).

See also