Check for Pareto-dominance

These functions take a numeric matrix as input where each column corresponds to a point and return a logical vector. The \(i\)th position of the latter is TRUE if the \(i\)th point is dominated by at least one other point for dominated. Likewise, the \(i\)th entry of nondominated is TRUE if it is not dominated by any other point.

dominated(x)

nondominated(x)

Arguments

x
[matrix]
Numeric \((m \times n)\) matrix where \(m\) is the number of objectives and \(n\) is the number of points.

Value

Logical vector where the \(i\)th component is TRUE if x[i] is dominated or nondominated respectively and FALSE otherwise.

See also

Examples

x = matrix(c(1, 1, 2, 2, 1, 3), byrow = FALSE, ncol = 3L)
dominated(x)

#> [1] FALSE  TRUE  TRUE

nondominated(x)

#> [1]  TRUE FALSE FALSE