(Modified) (Inverted) Generational distance.

The generational distance (GD) measures the distance of a point set \(X = \{r_1, \ldots, r_{|X|}\}\), e.g., a Pareto-front approximation, to a reference set \(R = \{r_1, \ldots, r_{|R|}\}\). Then GD is defined as $$ GD_p(A, R) = \frac{1}{|X|} \left(\sum_{i=1}^{|X|} d_i^p\right)^{1/p} $$ where \(d_i\) is the Euclidean distance of point \(x_i \in X\) to its nearest neigbor point in \(R\). The inverted generational distance works the other way around, i.e., $$ IGD_p(A, R) = \frac{1}{|R|} \left(\sum_{i=1}^{|R|} \hat{d}_i^p\right)^{1/p} $$ where \(\hat{d}_i\) is the respective nearest neighbor distance of \(r_i\) to any point in \(X\). Put differently, \(IGD_p(A, R) = GD_p(R, A)\). Functions gd and igd calcute these versions.

Schütze et al. [2] proposed a slight modification: $$ GD_p(A, R) = \left(\frac{1}{|X|} \sum_{i=1}^{|X|} d_i^p\right)^{1/p} $$ where the average is taken before the power operation. \(IGD_p\) is apdated analogeously. This versions are calclated by gd and igd if argument modified is set TRUE.

Ishibushi et al. [3] proposed another modification which works on the formulation by Schütze et al. (see above). They modified the distance calculation: $$ GD_p^{+}(A, R) = \left(\frac{1}{|X|} \sum_{i=1}^{|X|} d^{+^p}_i\right)^{1/p} $$ where \(d_i^{+} = \max\{x_i, z_i\}\). This version can be calculated with function gdp (the trailing p stands for “plus”).

Eventuelly, function ahd calculates the Average Hausdorff distance [2] which combines GD and IGD and is defined as $$ \Delta_p(A, R) = \max\{GD_p(A, R), IGD_p(A, R)\}. $$ By default, ahd uses the modified versions of \(GD\) and \(IGD\) respectively (see argument modified).

gd(x, y, p = 2, modified = TRUE, ...)

igd(x, y, p = 2, modified = TRUE, ...)

gdp(x, y, p = 2, modified = TRUE, ...)

igdp(x, y, p = 2, modified = TRUE, ...)

ahd(x, y, p = 2, modified = TRUE, ...)

Arguments

x
[matrix]
Point set in column major format.

y
[matrix]
Reference point set in column major format.

p
[numeric(1)]
Parameter \(p\) (see description).

modified
[logical(1)]
Should the modified GD/IGD calculation by Schuetze et al. [2] be used? Default is TRUE.

...
[any]
Not used.

Value

[numeric(1)] Scalar indicator value.

References

[1] David A. Van Veldhuizen and David A. Van Veldhuizen. Multiobjective evolutionary algorithms: classifications, analyses, and new innovations. Technical Report, Evolutionary Computation, 1999.

[2] Schütze, O., Esquivel, X.,Lara,A. ,Coello, C.A.C.: Using the averaged Hausdorff distance as a performance measure in evolutionary multiobjective optimization. IEEE Transactions on Evolutionary Computation 16, 504–522 (2012).

[3] Hisao Ishibuchi, Hiroyuki Masuda, Yuki Tanigaki, and Yusuke Nojima. Modified distance calculation in generational distance and inverted generational distance. In António Gaspar-Cunha, Carlos Henggeler Antunes, and Carlos Coello Coello, editors, Evolutionary Multi-Criterion Optimization, 110–125. Cham, 2015. Springer International Publishing.

See also