Dominated hypervolume (contribution)

The hypervolume (HV) indicator is arguebly one of the most often used performance indicator likely due to its straight-forward definition. Given a set of points \(X = \{x_1, \ldots, x_{|X|}\}\) and an anti-optimal reference point \(r \in R^m\) the HV-indicator, also called the S-metric, is defined as $$ HV(X, r) = \lambda_m\left(\bigcup_{x \in X} [x \preceq x' \preceq r]\right). $$ Here, \(\lambda_m\) is the \(m\)-dimensional Lebesgue measure. Informally, the hypervolume indicator is the space/volume enclosed by the point set and the anti-optimal (i.e., it is dominated by every point \(x \in X\)) reference point. It is known to be strictly monotonic. I.e., if point set \(X\) strictly dominates another point set \(Y\), then \(HV(X,r) > HV(Y, r)\) holds.

Function hv computes the dominated hypervolume of a set of points given a reference set whereby hv_contr computes the hypervolume contribution of each point.

If no reference point is given the nadir point of the set x is determined and a positive offset with default 1 is added. This is to ensure that the reference point is dominated by all of the points in the reference set (anti-optimality property).

hv(x, r = NULL, offset = 1, ...)

hv_contr(x, r = NULL, offset = 1)

Arguments

x
[matrix]
Matrix of points (column-wise).

r
[numeric | NULL]
Reference point. Set to the maximum in each dimension by default if not provided.

offset
[numeric(1)]
Offset to be added to each component of the reference point only in the case where no reference is provided and one is calculated automatically. Default is 1.

...
[any]
Not used at the moment.

Value

[numeric(1)] Dominated hypervolume in the case of hv and the dominated hypervolume contributions for each point in the case of hv_contr.

Note

Keep in mind that this function assumes all objectives to be minimized. In case at least one objective is to be maximized the matrix x needs to be transformed accordingly in advance.

See also