Heatmap

Visualize a Pareto-front approximation set by means of a heatmap. Here, on the \(x\)-axis the objectives are given in the order provided by obj.cols. The \(y\)-axis shows the different solutions. The color of the cells indicates the objective value of the respective (objective, solution)-pair. Works with any kind of objectives.

plot_heatmap(df, obj.cols = c("y1", "y2"))

Arguments

df
[data.frame]
Data.frame with columns at least obj.cols, “problem” and “algorithm”.

obj.cols
[character(>= 2)]
Column names of the objective function values. Default is c("y1", "y2").

Value

A ggplot object.

References

[1] T. Tušar and B. Filipič, Visualization of Pareto Front Approximations in Evolutionary Multiobjective Optimization: A Critical Review and the Prosection Method, in IEEE Transactions on Evolutionary Computation, vol. 19, no. 2, pp. 225-245, April 2015, doi: 10.1109/TEVC.2014.2313407.

See also