3D scatter plot

Given a data frame with the results of (multiple) runs of (multiple) different three-objective optimization algorithms on (multiple) problem instances the function generates 3D scatterplots of the obtained Pareto-front approximations.

plot_scatter3d(
  df,
  obj.cols = c("y1", "y2", "y3"),
  max.in.row = 4L,
  package = "scatterplot3d",
  ...
)

Arguments

df
[data.frame]
Data.frame with columns at least obj.cols, “problem” and “algorithm”.

obj.cols
[character(>= 3)]
Column names of the objective functions. Default is c("f1", "f2", "f3").

max.in.row
[integer(1)]
Maximum number of plots to be displayed side by side in a row. Default is 4.

package
[character(1L)]
Which package to use for 3d scatterplot generation? Possible choices are “scatterplot3d” (scatterplot3d), “plot3D”, (plot3D), “plot3Drgl” (plot3Drgl) or “plotly” (plotly). Default is “scatterplot3d”.

...
[any]
Further arguments passed down to scatterplot function.

Value

Nothing

References

[1] T. Tušar and B. Filipič, "Visualization of Pareto Front Approximations in Evolutionary Multiobjective Optimization: A Critical Review and the Prosection Method," in IEEE Transactions on Evolutionary Computation, vol. 19, no. 2, pp. 225-245, April 2015, doi: 10.1109/TEVC.2014.2313407

See also