Riesz \(s\)-energy

The Riesz \(s\)-energy is designed as a measure for the evenness of a set of points \(X = \{x_1, \ldots, x_{|X|}\}\). It is formally defined as $$ R_s(X) = \sum_{x \in X} \sum_{y \in X, y \neq x} k_s(x, y) $$ where function $$ k_s(x, y) = d(x, y)^{-s}, s > 0 $$ and $$ k_s(x, y) = -\log(d(x, y)), s = 0 $$ is the so-called Riesz \(s\)-kernel and \(d(x,y)\) is the Euclidean distance between \(x\) and \(y\). The parameter \(s \geq 0\) steers the desired degree of uniformity of the distribution with increasing emphasis on uniformity for \(s \to \infty\). See [1] for an application of the Riesz \(s\)-energy in multi-objective evolutionary optimization and [2, 3] for a mathematically rigorous introduction into the general idea.

rse(x, s, ...)

Arguments

x
[matrix]
Point set in column major format.

s
[numeric]
Degree of uniformity \(s \geq 0\) (see description).

...
[any]
Not used.

Value

Single numeric indicator value.

References

[1] J. G. Falcón-Cardona, H. Ishibuchi and C. A. C. Coello, Riesz s-energy-based Reference Sets for Multi-Objective optimization," 2020 IEEE Congress on Evolutionary Computation (CEC), 2020, pp. 1-8, doi: 10.1109/CEC48606.2020.9185833.

[2] D.P. Hardinand and E.B. Saff, Minimal Riesz energy point configurations for rectifiable \(d\)-dimensional manifolds, Advances in Mathematics, vol. 193, no. 1, pp. 174–204, 2005.

[3] D. P. Hardin and E. B. Saff, Discretizing Manifolds via Minimum Energy Points, Notices of the AMS, vol. 51, no. 10, pp. 1186–1194, 2004.

See also